requirements.txt
gradio==4.29.0
openai-whisper
transformers==4.41.1
torch>=1.10
librosa==0.10.1
moviepy==1.0.3
noisereduce==2.0.1
pysrt==1.1.2
ffmpeg-python==0.2.0
soundfile==0.12.1
scipy==1.12.0
numpy==1.26.4